// src/database/models/Permission.ts
export enum Permission {
    CREATE_EVENT = "CREATE_EVENT",
    EDIT_EVENT = "EDIT_EVENT",
    DELETE_EVENT = "DELETE_EVENT",

    CREATE_TRIP = "CREATE_TRIP",
    EDIT_TRIP = "EDIT_TRIP",
    DELETE_TRIP = "DELETE_TRIP",

    CREATE_COMMENT = "CREATE_COMMENT",
    EDIT_COMMENT = "EDIT_COMMENT",
    DELETE_COMMENT = "DELETE_COMMENT",

    BAN_USER = "BAN_USER",
    DELETE_USER = "DELETE_USER",
}